Change downloadExpiry and downloadLimit properties type from Int to Int64

+++ WooCommerce/Classes/ViewRelated/Products/Edit Product/EditableProductModel.swift
@@ -136,11 +136,11 @@ extension EditableProductModel: ProductFormDataModel, TaxClassRequestable {
     }
 
     var downloadLimit: Int64 {
-        Int64(product.downloadLimit)
+        product.downloadLimit
     }
 
     var downloadExpiry: Int64 {
-        Int64(product.downloadExpiry)
+        product.downloadExpiry
     }
 
     func isStockStatusEnabled() -> Bool {
